change all args to system_clock to integer(kind=8)
To give better granularity and mostly to avoid wrapping on long runs

+++ src/2d/amr2.f90
@@ -113,7 +113,7 @@ program amr2
     ! Timing variables
     integer ::  ttotal
     real(kind=8) ::ttotalcpu, cpu_start,cpu_finish 
-    integer :: clock_start, clock_finish, clock_rate
+    integer(kind=8) :: clock_start, clock_finish, clock_rate
     integer, parameter :: timing_unit = 48
     character(len=512) :: timing_line, timing_substr
     character(len=*), parameter :: timing_base_name = "timing."
